chore: moved AddressList to a separate file and removed votes functionality
